feat: move scaffold & check clap structs to cli crate
