+++ crates/bulloak/Cargo.toml
@@ -14,6 +14,20 @@ categories.workspace = true
 
 [dependencies]
 bulloak-core = { path = "../core" }
+anyhow = "1.0.75"
+clap = { version = "4.3.19", features = ["derive"] }
+figment = "0.10.19"
+forge-fmt = "0.2.0"
+indexmap = "2.0.0"
+once_cell = "1.18.0"
+owo-colors = "3.5.0"
+regex = "1.10.2"
+serde = "1.0.203"
+solang-parser = "0.3.2"
+thiserror = "1.0.61"
+unicode-xid = "0.2.4"
+# Security patch
+zerocopy = "0.7.31"
 
 [dev-dependencies]
 criterion = "0.5.1"
