+++ crates/bulloak/src/constants.rs
@@ -1,0 +1,2 @@
+// TODO: Consolidate this constant with INTERNAL_DEFAULT_SOL_VERSION.
+pub const DEFAULT_SOL_VERSION: &str = "0.8.0";
