+++ crates/bulloak/src/main.rs
@@ -1,8 +1,13 @@
 #![doc = include_str!("../README.md")]
 use std::process;
 
+mod check;
+mod cli;
+mod constants;
+mod scaffold;
+
 fn main() {
-    if let Err(e) = bulloak_core::config::run() {
+    if let Err(e) = crate::cli::run() {
         eprintln!("Error: {e:?}");
         process::exit(1);
     }
