+++ crates/core/src/check/utils.rs
@@ -1,8 +1,7 @@
+//! Check module utilities.
+
 /// Converts the start offset of a `Loc` to `(line, col)`. Modified from <https://github.com/foundry-rs/foundry/blob/45b9dccdc8584fb5fbf55eb190a880d4e3b0753f/fmt/src/helpers.rs#L54-L70>
-pub(crate) fn offset_to_line_column(
-    content: &str,
-    start: usize,
-) -> (usize, usize) {
+pub fn offset_to_line_column(content: &str, start: usize) -> (usize, usize) {
     debug_assert!(content.len() > start);
 
     // first line is `1`
@@ -19,6 +18,7 @@ pub(crate) fn offset_to_line_column(
     unreachable!("content.len() > start")
 }
 
-pub(crate) fn offset_to_line(content: &str, start: usize) -> usize {
+/// Returns the line where a byte offset is found.
+pub fn offset_to_line(content: &str, start: usize) -> usize {
     offset_to_line_column(content, start).0
 }
