+++ crates/core/src/hir/combiner.rs
@@ -274,8 +274,8 @@ mod tests {
         let mut discoverer = modifiers::ModifierDiscoverer::new();
         let modifiers = discoverer.discover(&ast);
 
-        let cfg: Config = Default::default();
-        let cfg = cfg.with_vm_skip(true);
+        let mut cfg: Config = Config::default();
+        cfg.emit_vm_skip = true;
         Ok(hir::translator::Translator::new().translate(&ast, modifiers, &cfg))
     }
