+++ crates/core/src/sol/mod.rs
@@ -16,9 +16,7 @@ pub(crate) use translator::Translator;
 
 /// Searches for and returns the first `ContractDefinition` found in a given
 /// `SourceUnit`.
-pub(crate) fn find_contract(
-    pt: &SourceUnit,
-) -> Option<Box<ContractDefinition>> {
+pub fn find_contract(pt: &SourceUnit) -> Option<Box<ContractDefinition>> {
     pt.0.iter().find_map(|part| match part {
         SourceUnitPart::ContractDefinition(contract) => Some(contract.clone()),
         _ => None,
